+++ tests/testthat/test-error-handler.R
@@ -9,9 +9,7 @@
 prepare_app <- function(app) {
   private <- app$.__enclos_env__$private
 
-  private$.routes <- app$get_routes()
-  private$.middleware <- app$get_middleware()
-  private$.params <- app$get_params()
+  private$.compile()
   invisible(NULL)
 }